Products

+++ database/migrations/2019_10_01_213421_create_models_categories_table.php
@@ -18,7 +18,7 @@ class CreateModelsCategoriesTable extends Migration
             $table->string('name');
             $table->string('slug')->unique();
             $table->text('description')->nullable();
-            $table->unsignedInteger('parent_id')->default(1)->nullable();
+            $table->unsignedBigInteger('parent_id')->default(1)->nullable();
             $table->boolean('featured')->default(0);
             $table->boolean('menu')->default(1);
             $table->string('image')->nullable();
